Cleaning some code
