+++ app/src/store/contexts/categoryContext.js
@@ -1,23 +1,21 @@
-import React, { createContext, useReducer, useEffect } from "react";
-import categoryReducer from "../reducers/categoryReducer";
+import React, { createContext, useState, useEffect } from "react";
 import Category from "../../services/Category";
-import actions from "../actions";
 
 export const CategoryContext = createContext();
 
 export default function CategoryContextProvider(props) {
-  const [categories, dispatch] = useReducer(categoryReducer, []);
+  const [categories, setCategories] = useState([]);
 
   useEffect(() => {
     async function fetchCategories() {
-      const categories = await Category.getAllCategories();
-      dispatch({ type: actions.category.getCategories, categories });
+      const categoriesArray = await Category.getAllCategories();
+      setCategories(categoriesArray);
     }
     fetchCategories();
   }, []);
 
   return (
-    <CategoryContext.Provider value={{ categories, dispatch }}>
+    <CategoryContext.Provider value={{ categories }}>
       {props.children}
     </CategoryContext.Provider>
   );
